Ensure a full parse when getting the spine parser.
If we're already going to traverse the entire snapshot, might as well do a proper parse as it's most likely already requested by other services or will be requested soon anyway. Fast forwarding a throw-away spine parser is wasted effort. Better pay a bit more once but have a parse tree ready for everyone to use.